put version code

+++ app/src/main/java/cy/agorise/crystalwallet/apigenerator/insightapi/GetEstimateFee.java
@@ -35,7 +35,7 @@ public abstract class GetEstimateFee {
             call.enqueue(new Callback<JsonObject>() {
                 @Override
                 public void onResponse(Call<JsonObject> call, Response<JsonObject> response) {
-                    listener.estimateFee((long) (response.body().get("2").getAsDouble()));
+                    listener.estimateFee((double) (response.body().get("2").getAsDouble()));
 
                 }
 
@@ -52,7 +52,7 @@ public abstract class GetEstimateFee {
     }
 
     public static interface estimateFeeListener{
-        public void estimateFee(long value);
+        public void estimateFee(double value);
         public void fail();
     }
